+++ core/shared/src/main/scala/fs2/Stream.scala
@@ -146,6 +146,9 @@ final class Stream[+F[_],+O] private (private val coreRef: Stream.CoreRef[F,O]) 
   /** Alias for `self through [[pipe.forall]]`. */
   def forall(f: O => Boolean): Stream[F, Boolean] = self through pipe.forall(f)
 
+  /** Alias for `self through [[pipe.groupBy]]`. */
+  def groupBy[O2](f: O => O2): Stream[F, (O2, Vector[O])] = self through pipe.groupBy(f)
+
   def interleave[F2[_],O2>:O](s2: Stream[F2,O2])(implicit R:RealSupertype[O,O2], S:Sub1[F,F2]): Stream[F2,O2] =
     (self through2v s2)(pipe2.interleave)